Add deck sidebar layout with project rail
- Add ProjectRail for project selection alongside sessions, files, and changes
- Move git changes from DiffPane into a sidebar GitChangesPanel in deck layout
- Support switching between classic and deck sidebar layouts
