+++ src/lib/workspaceTabGroups.test.ts
@@ -2,6 +2,7 @@ import { describe, expect, it } from "vitest";
 import { newTab, type WorkspaceTab } from "./layout";
 import type { Session } from "./session";
 import {
+  findTabForProject,
   replaceGroupInTabOrder,
   workspaceTabProject,
 } from "./workspaceTabGroups";
@@ -27,6 +28,23 @@ describe("workspaceTabProject", () => {
     const workspace = tab("t1", "s1");
     const sessions = [session("s1", "/Users/me/agent-terminal")];
     expect(workspaceTabProject(workspace, sessions)).toBe("agent-terminal");
+  });
+});
+
+describe("findTabForProject", () => {
+  it("matches a tab by project path, ignoring trailing slashes", () => {
+    const tabs = [tab("t1", "s1"), tab("t2", "s2")];
+    const sessions = [
+      session("s1", "/tmp/alpha"),
+      session("s2", "/tmp/beta"),
+    ];
+    expect(findTabForProject(tabs, sessions, "/tmp/beta/")?.id).toBe("t2");
+  });
+
+  it("returns undefined when no open tab belongs to the project", () => {
+    const tabs = [tab("t1", "s1")];
+    const sessions = [session("s1", "/tmp/alpha")];
+    expect(findTabForProject(tabs, sessions, "/tmp/beta")).toBeUndefined();
   });
 });
