+++ src/lib/workspaceTabGroups.ts
@@ -4,27 +4,43 @@ import {
   type WorkspaceTab,
 } from "./layout";
 import { projectName } from "./paths";
+import { sameProjectPath } from "./recents";
 import type { Session } from "./session";
 
-export function workspaceTabProject(
+export function workspaceTabCwd(
   tab: WorkspaceTab,
   sessions: Session[],
 ): string | null {
   for (const id of leafIds(tab.layout)) {
     const session = sessions.find((entry) => entry.id === id);
-    if (session) {
-      const name = projectName(session.cwd);
-      return name === "~" ? null : name;
-    }
+    if (session?.cwd && session.cwd !== "~") return session.cwd;
   }
 
   const file = focusedFileTab(tab);
-  if (file) {
-    const name = projectName(file.cwd);
-    return name === "~" ? null : name;
-  }
+  if (file?.cwd && file.cwd !== "~") return file.cwd;
 
   return null;
+}
+
+export function workspaceTabProject(
+  tab: WorkspaceTab,
+  sessions: Session[],
+): string | null {
+  const cwd = workspaceTabCwd(tab, sessions);
+  if (!cwd) return null;
+  const name = projectName(cwd);
+  return name === "~" ? null : name;
+}
+
+export function findTabForProject(
+  tabs: WorkspaceTab[],
+  sessions: Session[],
+  path: string,
+): WorkspaceTab | undefined {
+  return tabs.find((tab) => {
+    const cwd = workspaceTabCwd(tab, sessions);
+    return cwd ? sameProjectPath(cwd, path) : false;
+  });
 }
 
 export function isGroupableProject(
